+++ FRAuth/FRAuth/Callbacks/Binding.swift
@@ -1,0 +1,38 @@
+// 
+//  Binding.swift
+//  FRAuth
+//
+//  Copyright (c) 2022 ForgeRock. All rights reserved.
+//
+//  This software may be modified and distributed under the terms
+//  of the MIT license. See the LICENSE file for details.
+//
+
+
+import Foundation
+
+
+/// Device Binding protocol to provide utility methods for DeviceBindingCallback and DeviceSigningVerifierCallback
+protocol Binding {
+    
+    /// Create the interface for the authentication type
+    /// - Parameter type: The Device Binding Authentication Type
+    /// - Returns: The recommended  DeviceAuthenticator that can handle the provided DeviceBindingAuthenticationType
+    func getDeviceAuthenticator(type: DeviceBindingAuthenticationType) -> DeviceAuthenticator
+    
+    /// Get Expiration date for the signed token, claim "exp" will be set to the JWS.
+    /// - Returns: The expiration date
+    func getExpiration(timeout: Int?) -> Date
+}
+
+
+extension Binding {
+    func getDeviceAuthenticator(type: DeviceBindingAuthenticationType) -> DeviceAuthenticator {
+        return type.getAuthType()
+    }
+    
+    
+    func getExpiration(timeout: Int?) -> Date {
+        return Date().addingTimeInterval(Double(timeout ?? 60))
+    }
+}