+++ FRCore/FRCoreTests/FRCore/Authenticator/CryptoKeyTests.swift
@@ -1,6 +1,6 @@
 // 
-//  KeyAwareTests.swift
-//  FRAuthTests
+//  CryptoKeyTests.swift
+//  FRCoreTests
 //
 //  Copyright (c) 2022 ForgeRock. All rights reserved.
 //
@@ -10,23 +10,23 @@
 
 
 import XCTest
-@testable import FRAuth
+@testable import FRCore
 
 
-class KeyAwareTests: XCTestCase {
+class CryptoKeyTests: XCTestCase {
     
     func test_01_basic_init() {
         let userId = "Test User Id 1"
-        let keyAware = KeyAware(userId: userId)
-        XCTAssertEqual(keyAware.timeout, 60)
+        let cryptoKey = CryptoKey(keyId: userId)
+        XCTAssertEqual(cryptoKey.timeout, 60)
     }
     
     
     func test_02_keyBuilderQuery() {
         let userId = "Test User Id 2"
-        let key = KeyAware.getKeyAlias(keyName: userId)
-        let keyAware = KeyAware(userId: userId)
-        let query = keyAware.keyBuilderQuery()
+        let key = CryptoKey.getKeyAlias(keyName: userId)
+        let cryptoKey = CryptoKey(keyId: userId)
+        let query = cryptoKey.keyBuilderQuery()
         
         XCTAssertEqual(query[String(kSecAttrKeyType)] as! String, String(kSecAttrKeyTypeECSECPrimeRandom))
         XCTAssertEqual(query[String(kSecAttrKeySizeInBits)] as! Int, 256)
@@ -43,60 +43,60 @@ class KeyAwareTests: XCTestCase {
     
     func test_03_createKeyPair() {
         let userId = "Test User Id 3"
-        let key = KeyAware.getKeyAlias(keyName: userId)
-        let keyAware = KeyAware(userId: userId)
-        let query = keyAware.keyBuilderQuery()
+        let key = CryptoKey.getKeyAlias(keyName: userId)
+        let cryptoKey = CryptoKey(keyId: userId)
+        let query = cryptoKey.keyBuilderQuery()
         var privateKey: SecKey?
         
         do {
-            privateKey = KeyAware.getSecureKey(keyAlias: key)
+            privateKey = CryptoKey.getSecureKey(keyAlias: key)
             XCTAssertNil(privateKey)
             
-            let keyPair = try keyAware.createKeyPair(builderQuery: query)
+            let keyPair = try cryptoKey.createKeyPair(builderQuery: query)
             XCTAssertEqual(keyPair.keyAlias, key)
             
-            privateKey = KeyAware.getSecureKey(keyAlias: key)
+            privateKey = CryptoKey.getSecureKey(keyAlias: key)
             XCTAssertNotNil(privateKey)
         } catch {
             XCTFail("Failed to create KeyPair")
         }
-        KeyAware.deleteKey(keyAlias: key)
+        CryptoKey.deleteKey(keyAlias: key)
     }
     
     
     func test_04_getSecureKey() {
         let userId = "Test User Id 4"
-        let key = KeyAware.getKeyAlias(keyName: userId)
-        let keyAware = KeyAware(userId: userId)
-        let query = keyAware.keyBuilderQuery()
+        let key = CryptoKey.getKeyAlias(keyName: userId)
+        let cryptoKey = CryptoKey(keyId: userId)
+        let query = cryptoKey.keyBuilderQuery()
         var privateKey: SecKey?
         
-        privateKey = KeyAware.getSecureKey(keyAlias: key)
+        privateKey = CryptoKey.getSecureKey(keyAlias: key)
         XCTAssertNil(privateKey)
         
         SecKeyCreateRandomKey(query as CFDictionary, nil)
         
-        privateKey = KeyAware.getSecureKey(keyAlias: key)
+        privateKey = CryptoKey.getSecureKey(keyAlias: key)
         XCTAssertNotNil(privateKey)
         
-        KeyAware.deleteKey(keyAlias: key)
+        CryptoKey.deleteKey(keyAlias: key)
     }
     
     
     func test_05_deleteKey() {
         let userId = "Test User Id 5"
-        let key = KeyAware.getKeyAlias(keyName: userId)
-        let keyAware = KeyAware(userId: userId)
-        let query = keyAware.keyBuilderQuery()
+        let key = CryptoKey.getKeyAlias(keyName: userId)
+        let cryptoKey = CryptoKey(keyId: userId)
+        let query = cryptoKey.keyBuilderQuery()
         var privateKey: SecKey?
         
         SecKeyCreateRandomKey(query as CFDictionary, nil)
         
-        privateKey = KeyAware.getSecureKey(keyAlias: key)
+        privateKey = CryptoKey.getSecureKey(keyAlias: key)
         XCTAssertNotNil(privateKey)
         
-        KeyAware.deleteKey(keyAlias: key)
-        privateKey = KeyAware.getSecureKey(keyAlias: key)
+        CryptoKey.deleteKey(keyAlias: key)
+        privateKey = CryptoKey.getSecureKey(keyAlias: key)
         XCTAssertNil(privateKey)
     }
     
@@ -105,7 +105,7 @@ class KeyAwareTests: XCTestCase {
         let userId1 = "Test User Id 6-1"
         let userId2 = "Test User Id 6-2"
         
-        XCTAssertEqual(KeyAware.getKeyAlias(keyName: userId1), KeyAware.getKeyAlias(keyName: userId1))
-        XCTAssertNotEqual(KeyAware.getKeyAlias(keyName: userId1), KeyAware.getKeyAlias(keyName: userId2))
+        XCTAssertEqual(CryptoKey.getKeyAlias(keyName: userId1), CryptoKey.getKeyAlias(keyName: userId1))
+        XCTAssertNotEqual(CryptoKey.getKeyAlias(keyName: userId1), CryptoKey.getKeyAlias(keyName: userId2))
     }
 }